move definitions in env
